Fix relocation of map_fd
The code from https://github.com/safchain/goebpf expects each map to be
declared in its own ELF section, like this:
> struct bpf_map_def SEC("maps/tcp_event") tcp_event = {

This patch switches the kernel code to this model.

Also, this adds support for iterating over a "maps/*" section.

Also, improve error messages when a map variable is used but not found
in the "maps/*" section.