Fix polling on perf fds
POLLIN events are not generated by default when an record is added in
the perf ring buffer. perf_event_open() must enable that feature
explicitely with "attr.wakeup_events = 1".

Since "wakeup_events" is declared in a C union, this is not available
from Golang. So we do it in a C function.