Guess offsets
Instead of requiring a different ebpf object file per kernel version /
distro, we try to guess the offsets of the fields we access in the eBPF
program.

We do so by creating connections to a server we start, trying to read
through struct sock, and comparing the obtained values with the expected
ones.