Set SO_LINGER to 0
We make a lot of connections when we get the offsets and they can
potentially fill the conntrack table because the state the connections
remain in when they get closed is TIME_WAIT and it stays like that for a
while (around 2 minutes).

We set SO_LINGER to 0 so the connection state after closing is CLOSE
instead. In this way, they will disappear from the conntrack table after
around 10 seconds.