Fix guessing the source port
The source port appears twice in the kernel structure, one in host
endianness in (struct sock_common).skc_num and another one in network
endianness in (struct inet_sock)->inet_sport.

We were finding the first one, but that's not the one we should use (it
gets reset to 0 in some scenarios).

Search the source port in the correct endianness to find the second one.
To avoid returning the first one if it happens that the source port used
is the same in both endiannesses, we start guessing at the family
offset.